refactor(): 패키지 구조 변경 및 필요없는 클래스 삭제
