+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,7 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
-
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
