--- src/test/java/chess/domain/ChessBoardTest.java
@@ -1,14 +1,0 @@
-package chess.domain;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class ChessBoardTest {
-//    @DisplayName("체스 보드 객체 생성 확인")
-//    @Test
-//    void 체스_보드_객체_생성() {
-//        assertThat(ChessBoard.getChessBoard().size()).isEqualTo(64);
-//    }
-}
