+++ src/test/java/chess/domain/ChessGameTest.java
@@ -15,9 +15,4 @@ public class ChessGameTest {
 
         assertThat(chessGame.getCurrentPieces()).isInstanceOf(CurrentPieces.class);
     }
-
-//    @DisplayName("source 위치 ")
-//    @Test
-//    void name() {
-//    }
 }
