+++ src/test/java/chess/domain/piece/EmptyTest.java
@@ -1,6 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
