+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -1,6 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -55,12 +54,29 @@ public class PawnTest {
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
-    @DisplayName("Pawn 규칙에 따라 이미 움직인 Pawn을 움직이는 경우 예외 - 2칸 이동 ")
+    @DisplayName("Pawn 규칙에 따라 이미 움직인 Pawn 을 움직이는 경우 예외 - 2칸 이동 ")
     @Test
     void pawn_이미_이동_2칸_예외() {
         Pawn pawn = new Pawn(Position.of('a', '6'), "P", Color.BLACK);
 
         assertThatThrownBy(() -> pawn.move(Position.of('a', '4'), currentPieces))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("뒤로 이동할 경우 예외")
+    @Test
+    void pawn_뒤로_이동() {
+        List<Piece> current = Arrays.asList(
+                new Pawn(Position.of('a', '7'), "P", Color.BLACK),
+                new Pawn(Position.of('a', '2'), "P", Color.WHITE));
+        CurrentPieces currentPieces = new CurrentPieces(current);
+
+        Piece pawn1 = currentPieces.findByPosition(Position.of('a', '7'));
+        Piece pawn2 = currentPieces.findByPosition(Position.of('a', '2'));
+
+        assertThatThrownBy(() -> pawn1.move(Position.of('a', '8'), currentPieces))
+                .isInstanceOf(IllegalArgumentException.class);
+        assertThatThrownBy(() -> pawn2.move(Position.of('a', '1'), currentPieces))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -96,20 +112,6 @@ public class PawnTest {
     }
 
     @Test
-    void 검은말이_상대편_말을_공격한다_예외() {
-        List<Piece> current = Arrays.asList(
-                new Pawn(Position.of('a', '7'), "P", Color.BLACK));
-        CurrentPieces currentPieces = new CurrentPieces(current);
-
-        Position source = Position.of('a', '7'); // 비숍 위치
-        Position target = Position.of('b', '6'); // 옮기고자 하는 위치
-        Piece pawn = currentPieces.findByPosition(source);
-
-        assertThatThrownBy(() -> pawn.move(target, currentPieces))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
     void 하얀말이_상대편_말을_공격한다() {
         List<Piece> current = Arrays.asList(
                 new Pawn(Position.of('a', '2'), "p", Color.WHITE),
@@ -123,19 +125,5 @@ public class PawnTest {
         pawn.move(target, currentPieces);
 
         assertThat(currentPieces.getCurrentPieces().size()).isEqualTo(1);
-    }
-
-    @Test
-    void 하얀말이_상대편_말을_공격한다_예외() {
-        List<Piece> current = Arrays.asList(
-                new Pawn(Position.of('a', '2'), "p", Color.WHITE));
-        CurrentPieces currentPieces = new CurrentPieces(current);
-
-        Position source = Position.of('a', '2'); // 비숍 위치
-        Position target = Position.of('b', '3'); // 옮기고자 하는 위치
-        Piece pawn = currentPieces.findByPosition(source);
-
-        assertThatThrownBy(() -> pawn.move(target, currentPieces))
-                .isInstanceOf(IllegalArgumentException.class);
     }
 }
